+++ src/otx/algorithms/detection/adapters/mmdet/apis/train.py
@@ -147,7 +147,7 @@ def train_detector(model, dataset, cfg, distributed=False, validate=False, times
         else:
             dtype = torch.float32
         model.train()
-        model, optimizer = torch.xpu.optimize(model, optimizer=optimizer, dtype=dtype)
+        model, optimizer = torch.xpu.optimize(model, optimizer=optimizer)
 
     runner = build_runner(
         cfg.runner, default_args=dict(model=model, optimizer=optimizer, work_dir=cfg.work_dir, logger=logger, meta=meta)
